% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_diarieducacio.R
\name{theme_diarieducacio}
\alias{theme_diarieducacio}
\title{Diari Educació theme}
\usage{
theme_diarieducacio(base_size = 12, base_family = "Open Sans")
}
\arguments{
\item{base_size}{Base font and elements size (Default is 12)}

\item{base_family}{Base font family (Default is "Open Sans" - may be needed to install with library(extrafont) font_import(pattern="OpenSans"))}
}
\description{
This function customises a theme for Diari Educació charts
}
\examples{
theme_diarieducacio()
}
\keyword{theme}
